pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openclaw-jobradar"
version = "0.1.0"
description = "AI-powered job search agent for Germany-based AI/tech roles"
readme = "README.md"
requires-python = ">=3.11"
license = "GPL-3.0-only"
dependencies = [
    "openai>=1.30.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0",
    "python-jobspy>=1.1.0",
    "openpyxl>=3.1.0",
    "rapidfuzz>=3.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    # CV format support
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.5",
    "aiofiles>=23.2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4.0",
]
mcp = [
    "mcp>=0.1.0",
]

[project.scripts]
jobradar = "src.main:main"
jobradar-web = "src.web.app:create_app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
